requirements.txt
click==8.3.1
colorama==0.4.6
gutenberg-cleaner==0.1.6
joblib==1.5.3
nltk==3.9.4
pip==25.3
python-dotenv==1.2.2
regex==2026.2.28
tqdm==4.67.3 